Works brilliantly, with caveats
Cannot compile the entire swissprot in one go, as the data created by ~547,000 proteins in the form of json data files, ted_IDs and list iterating goes way over the available RAM (32GB). for users, it is recommended to divide databases into smaller chunks for loading TED tiles, at a maximum of a gigabyte per chunk.